discord_app/pyproject.toml
[tool.poetry]
name = "discord_app"
version = "1.0"
description = ""
authors = ["lesteenman <lesteenman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2"
#pynamodb = "^5.2.1"
loguru = "^0.6.0"
#deprecation = "^2.1.0"
#aws-xray-sdk = "^2.10.0"
PyNaCl = "^1.5.0"

[tool.poetry.dev-dependencies]
boto3 = "^1.26.6"
flake8 = "^5.0"
pytest = "^7.1.1"
pytest-mock = "^3.5.1"
pytest-asyncio = "^0.18.3"
localstack = "^0.14.2"
pytest-env = "^0.6.2"
docker = "^5.0.3"
opnieuw = "^1.1.0"
pytest-cov = "^3.0.0"
coverage = {extras = ["toml"], version = "^6.3"}
moto = "^3.1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]
